requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==24.2
pypdf==5.4.0
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.9.0
